Fix improper location serialization; use "put" instead of "replace" when updating maps
When packing a location, I was storing a reference to the world instead of the world's name. This would have made all teams automatically invalid.

I misread the documentation when implementing the ConfigXYZ classes and thought that "put" would only set pairs if they hadn't been asigned yet and that "replace" would either "put" in a new value or "replace" a previous value. Both assumptions were incorrect.